+++ libft/includes/libft.h
@@ -6,7 +6,7 @@
 /*   By: nroman <marvin@42.fr>                      +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/02/21 19:34:51 by nroman            #+#    #+#             */
-/*   Updated: 2018/06/02 08:22:19 by nroman           ###   ########.fr       */
+/*   Updated: 2018/06/18 08:34:25 by nroman           ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -66,7 +66,7 @@ char				*ft_strsub(char const *s, unsigned int start, size_t len);
 char				*ft_strjoin(char const *s1, char const *s2);
 char				*ft_strtrim(char const *s);
 char				**ft_strsplit(char const *str, char c);
-char				*ft_itoa(int n);
+char				*ft_itoa(long long int n);
 char				*ft_itoa_base(long long int n, int base);
 void				ft_putchar(char c);
 void				ft_putstr(char const *str);
